+++ auto-file-and-function-comments/Python/tests/test_file.py
@@ -1,9 +1,12 @@
 def sum(a:int, b:int)->int:
     return a + b
 
-def difference(minuend:float, subtrahend:float)->float:
+def difference(minuend: float, subtrahend: float) -> float:
     dif = minuend - subtrahend
     return dif
 
-def printLine(line):
+def printLine(line:str):
     print(line, end="\n")
+
+def empty():
+    print("")