+++ auto-file-and-function-comments/CHANGELOG.md
@@ -1,7 +1,11 @@
 # Change Log
 
-All notable changes to the "auto-file-and-function-comments" extension will be documented in this file.
+All notable changes to the "Auto File and Function Comments" extension will be documented in this file.
 
 ## [1.0.0]
 
 - Initial release
+
+## [1.0.1]
+
+- Added functionality for python files.